Drop SDK dependency
If the OTLP exporter is to be the default used by the SDK, then it doesn't
sit right that the exporter would depend on the SDK. This commit makes it
so that the exporter doesn't directly depend on the SDK to function, which
means that the SDK can then recommend the exporter as an optional dependency
without getting into a dependency cycle.

This wasn't o difficult since the only place we were still using the SDK
directly was in the test suite. It would still be a good idea to have an
optional test that ensured the exporter worked with the real SDK classes,
but for now it's enough to break the hard dependency.