Allow encoding arbitrary data structures
While the attributes of some OpenTelemetry entities are limited
(like span attributes, which can only hold values that are nuclear
types, or lists of values that are themselves nuclear types), the
OTLP schema does not enforce this limitation. This is because not
_all_ entities are subject to the same limitations: in particular,
the attributes of log records in the Logs API are expressly allowed
to hold arbitrary data.

A change like this is necessary for eventually supporting logs and
metrics.